Refactor media module to smaller modules
