+++ js/core/core.mjs
@@ -1,3 +1,2 @@
 //Core app domains
 export * from './media/media.mjs';
-export * from './media/media_player.mjs';
