+++ js/core/media/media_metadata_restclient.mjs
@@ -1,0 +1,19 @@
+export class MediaMetaDataRestClient {
+    #baseUri;
+
+    constructor(baseUri) {
+        //TODO: validate URI...
+        this.#baseUri = baseUri;
+    }
+
+    async getById(id) {
+        let uri = this.#buildUri(`medias/${id}`);
+
+        return await fetch(uri)
+            .then(response => response.json());
+    }
+
+    #buildUri(resourcePath) {
+        return `${this.#baseUri}/${resourcePath}`;
+    }
+}
